+++ test/td_dd_web/controllers/system_controller_test.exs
@@ -353,7 +353,7 @@ defmodule TdDdWeb.SystemControllerTest do
 
       ElasticsearchMock
       |> expect(:request, fn
-        _, :post, "/structures/_search", %{from: 0, size: 1000, query: query}, [] ->
+        _, :post, "/structures/_search", %{from: 0, size: 1000, query: query}, _ ->
           assert query == %{
                    bool: %{
                      filter: [
@@ -402,7 +402,7 @@ defmodule TdDdWeb.SystemControllerTest do
 
   defp expect_search(results \\ nil) do
     ElasticsearchMock
-    |> expect(:request, fn _, :post, "/structures/_search", _, [] ->
+    |> expect(:request, fn _, :post, "/structures/_search", _, _ ->
       results
       |> List.wrap()
       |> SearchHelpers.hits_response()